% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_distance.R
\name{abc_distance}
\alias{abc_distance}
\title{abc_distance}
\usage{
abc_distance(summaries_parameters, summaries_weights, N)
}
\arguments{
\item{summaries_parameters}{vector of summaries}

\item{summaries_weights}{vector with weights of summaries computed from the data}

\item{N}{number of populations of JRNMMs}
}
\value{
ABC distance for the given summaries
}
\description{
Compute the distance given the summaries
}
