% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_summaries.r
\name{abc_summaries}
\alias{abc_summaries}
\title{abc_summaries}
\usage{
abc_summaries(N, summaries_X, summaries_Y, summaries_extra)
}
\arguments{
\item{N}{number of populations of JRNMMs}

\item{summaries_X}{summaries of observed dataset X}

\item{summaries_Y}{summaries of synthetic dataset X}

\item{summaries_extra}{vector of quantities needed to compute the summary statistics}
}
\value{
distance value among two sets of summaries
}
\description{
Computation of the three ABC summary statistics
}
