% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_summaries.r
\name{abc_summaries_extra}
\alias{abc_summaries_extra}
\title{abc_summaries_extra}
\usage{
abc_summaries_extra(X, T, h)
}
\arguments{
\item{X}{Nxn matrix of observed dataset/path of N populations of JR-NMMs with n points each}

\item{T}{time horizon for path simulation}

\item{h}{step size for path simulation}
}
\value{
vector of required quantities for then calculating the summary statistics
}
\description{
Determination of quantities needed to determine ABC summaries of a given dataset
}
\details{
#-------------------------------------------------------------------------
}
