% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_summaries.r
\name{abc_summaries_weights}
\alias{abc_summaries_weights}
\title{abc_summaries_weights}
\usage{
abc_summaries_weights(summaries, summaries_extra, N)
}
\arguments{
\item{summaries}{summary statistics of observed vs simulated dataset}

\item{summaries_extra}{vector of quantities needed to compute the summary statistics}

\item{N}{number of populations in the JRNMM}
}
\value{
vector of summary weights
}
\description{
Computation of weights (for distance measure among two sets of summaries) from observed data
}
