% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMCABCnJRNMM_acceptrate.R
\name{abcpilot}
\alias{abcpilot}
\title{abcpilot}
\usage{
abcpilot(
  data,
  Pr_cont,
  extra_model,
  numsim,
  whichprior = "unif",
  subsamplingby = 1
)
}
\arguments{
\item{data}{observed data}

\item{Pr_cont}{Matrix of lower/upper bounds for the uniform priors for the continuous parameters}

\item{extra_model}{List with Number of points to simulate N, final time T, grid, time step h, initial position, vector of diagonal entries of the gamma matrix, vector of diagonal entries of the sigma matrix, Theta0}

\item{numsim}{Number of simulation in the ABC pilot}

\item{whichprior}{Chosen prior. Here only 'unif' is available}

\item{subsamplingby}{every how many simulated points the observation should be taken. The default=1, i.e., no subsampling}
}
\description{
Reference-table ABC for time-varying thresholds computed as percentiles of ACCEPTED distances with a fixed budget (number of simulations)
}
