% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.r
\name{KmatrixgivenLc}
\alias{KmatrixgivenLc}
\title{Code to calculate the Kmatrix given L and c}
\usage{
KmatrixgivenLc(N, L, c)
}
\arguments{
\item{N}{number of populations of neurons}

\item{L}{strength parameter}

\item{c}{[0,1] parameter}
}
\value{
Kmatrix given L and c
}
\description{
Sampling from the Gaussian Perturbation kernel for the continuous parameters
}
