% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.r
\name{perturb_continuous}
\alias{perturb_continuous}
\title{perturb_continuous}
\usage{
perturb_continuous(Theta, Sigma)
}
\arguments{
\item{Theta}{Mean of the Gaussian perturbation kernel}

\item{Sigma}{covariance matrix of the Gaussian perturbation kernel}
}
\value{
1 sampled value from the Gaussian Perturbation kernel for the continuous parameters
}
\description{
Sampling from the Gaussian Perturbation kernel for the continuous parameters
}
