% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_kernels.R
\name{perturb_continuous_withinprior}
\alias{perturb_continuous_withinprior}
\title{perturb_continuous_withinprior}
\usage{
perturb_continuous_withinprior(Theta, Sigma, Pr_cont)
}
\arguments{
\item{Theta}{Mean of the Gaussian perturbation kernel}

\item{Sigma}{covariance matrix of the Gaussian perturbation kernel}

\item{Pr_cont}{Matrix of lower/upper bounds for the uniform priors for the continuous parameters}
}
\value{
1 sampled value from the Gaussian Perturbation kernel for the continuous parameters
}
\description{
Sampling from the Gaussian Perturbation kernel for the continuous parameters, repeated until the sampled values are within the prior region
}
