% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_prior.R
\name{problemprior}
\alias{problemprior}
\title{problemprior}
\usage{
problemprior(theta, draw, Pr_cont, whichprior = "unif")
}
\arguments{
\item{theta}{Parameter Vector;}

\item{draw}{Draw=0 if we want to compute the product of the pdfs of the priors
Draw =1 if we want to sample from the priors}

\item{Pr_cont}{Matrix of lower/upper bounds for the uniform priors for the continuous parameters}

\item{whichprior}{1 uniform, 2 lognormal, 3 exponential}
}
\value{
Simulation of observations from all models (for draw=1) or pdf of the priors (for draw=0), returned as vector;
}
\description{
Either pdf of priors (for draw=0) or sample from the prior (for draw =1)
}
