% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMCABCnJRNMM.r
\docType{package}
\name{SMCABCnJRNMM}
\alias{SMCABCnJRNMM-package}
\alias{SMCABCnJRNMM}
\title{SMCABCnJRNMM}
\description{
Inference via SMC-ABC for the a network of N stochastics JR-NMMs
}
\author{
Massimiliano Tamborrino
}
