% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMCABCnJRNMM_acceptrate.R
\name{SMCABCnJRNMM_acceptrate}
\alias{SMCABCnJRNMM_acceptrate}
\title{SMCABCnJRNMM_acceptrate}
\usage{
SMCABCnJRNMM_acceptrate(
  data,
  Pr_cont,
  stay_prob,
  extra_model,
  ABCthreshold,
  minaccept.rate,
  numparticles,
  alpha,
  sampling,
  attempt,
  folder,
  whichprior = "unif",
  subsamplingby = 1
)
}
\arguments{
\item{data}{observed data}

\item{Pr_cont}{Matrix of lower/upper bounds for the uniform priors for the continuous parameters}

\item{stay_prob}{Probability of not perturbing the sampled discrete connection probability}

\item{extra_model}{List with Number of points to simulate N, final time T, grid, time step h, initial position, vector of diagonal entries of the gamma matrix, vector of diagonal entries of the sigma matrix, Theta0}

\item{ABCthreshold}{Initial value for the threshold for the guided approach}

\item{minaccept.rate}{Lowest acceptance rate after which we stop the algorithm}

\item{numparticles}{number of particles to keep}

\item{alpha}{percentile to compute the ABC threshold out of all distances}

\item{sampling}{specify the chosen samplers. The options are standard or olcm}

\item{attempt}{number of iteration}

\item{folder}{specify the folder where the results are going to be saved in}

\item{whichprior}{choose between 'unif','lognormal' and 'exp'}

\item{subsamplingby}{every how many simulated points the observation should be taken. The default=1, i.e., no subsampling}
}
\description{
SMC-ABC for time-varying thresholds computed as percentiles of ACCEPTED distances with a fixed budget (number of simulations)
}
