% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMCABCnJRNMM_acceptrate.R
\name{SMCABCnJRNMM_r1}
\alias{SMCABCnJRNMM_r1}
\title{SMCABCnJRNMM_r1}
\usage{
SMCABCnJRNMM_r1(
  data,
  Pr_cont,
  extra_model,
  ABCthreshold,
  minaccept.rate,
  numparticles,
  attempt,
  whichprior = "unif",
  subsamplingby = 1
)
}
\arguments{
\item{data}{observed data}

\item{Pr_cont}{Matrix of lower/upper bounds for the uniform priors for the continuous parameters}

\item{extra_model}{List with Number of points to simulate N, final time T, grid, time step h, initial position, vector of diagonal entries of the gamma matrix, vector of diagonal entries of the sigma matrix, Theta0}

\item{ABCthreshold}{Initial value for the threshold for the guided approach}

\item{minaccept.rate}{Lowest acceptance rate after which we stop the algorithm}

\item{numparticles}{number of particles to keep}

\item{attempt}{number of iteration}

\item{whichprior}{choose between 'unif','lognormal' and and 'exp'}

\item{subsamplingby}{every how many simulated points the observation should be taken. The default=1, i.e., no subsampling}
}
\description{
SMC-ABC for time-varying thresholds computed as percentiles of ACCEPTED distances with a fixed budget (number of simulations)
}
